medical_etls/part1/requirements.txt
mysql-connector-python==8.0.21